Add the ability to sign in.
Improved the UI.
Added the link to the forgot password page and sign up page.
Added the sign in flow to the sign in activity.
Updated to Java 1.8.
Added simple instrumental test to check that the sign in and out is working through UI.
Updated dependancies to AndroidX + added testing dependancies.